Post-review fixes 1
